--- InputValidation.cs
@@ -1,83 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Globalization;
-using System.Threading.Tasks;
-using System.Text.RegularExpressions;
-
-namespace Tools {
-    public class InputValidation {
-        bool invalid = false;
-
-        public static bool checkLetters(string input) {
-            int count = Regex.Matches(input, @"[a-zA-Z\-\ ]").Count;
-
-            if (input.Length != 0 && count == input.Length) {
-                return true;
-            } else {
-                return false;
-            }
-        }
-
-        public static bool checkNumbers(string input) {
-            int count = Regex.Matches(input, @"[0-9]").Count;
-
-            if (input.Length != 0 && count == input.Length) {
-                return true;
-            } else {
-                return false;
-            }
-        }
-
-        public static bool checkDecimals(string input) {
-            int count = Regex.Matches(input, @"[0-9\.]").Count;
-
-            if (input.Length != 0 && count == input.Length) {
-                return true;
-            } else {
-                return false;
-            }
-        }
-
-        public bool IsValidEmail(string strIn) {
-            invalid = false;
-            if (String.IsNullOrEmpty(strIn))
-                return false;
-
-            // Use IdnMapping class to convert Unicode domain names. 
-            try {
-                strIn = Regex.Replace(strIn, @"(@)(.+)$", this.DomainMapper,
-                                      RegexOptions.None, TimeSpan.FromMilliseconds(200));
-            } catch (RegexMatchTimeoutException) {
-                return false;
-            }
-
-            if (invalid)
-                return false;
-
-            // Return true if strIn is in valid e-mail format. 
-            try {
-                return Regex.IsMatch(strIn,
-                      @"^(?("")("".+?(?<!\\)""@)|(([0-9a-z]((\.(?!\.))|[-!#\$%&'\*\+/=\?\^`\{\}\|~\w])*)(?<=[0-9a-z])@))" +
-                      @"(?(\[)(\[(\d{1,3}\.){3}\d{1,3}\])|(([0-9a-z][-\w]*[0-9a-z]*\.)+[a-z0-9][\-a-z0-9]{0,22}[a-z0-9]))$",
-                      RegexOptions.IgnoreCase, TimeSpan.FromMilliseconds(250));
-            } catch (RegexMatchTimeoutException) {
-                return false;
-            }
-        }
-
-        private string DomainMapper(Match match) {
-            // IdnMapping class with default property values.
-            IdnMapping idn = new IdnMapping();
-
-            string domainName = match.Groups[2].Value;
-            try {
-                domainName = idn.GetAscii(domainName);
-            } catch (ArgumentException) {
-                invalid = true;
-            }
-            return match.Groups[1].Value + domainName;
-        }
-    }
-}
